13/04 13:12

+++ HW1/signals.cpp
@@ -7,9 +7,9 @@
    Synopsis: handle the Control-C */
 #include "signals.h"
 extern int cur_pid;
+extern int smash_pid;
 
-
-void ctrl_c_smash_handler(int sig_num){
+/*void ctrl_c_smash_handler(int sig_num){
 	sigset_t mask_set;
 	sigset_t old_set;
 	sigfillset(&mask_set);
@@ -17,36 +17,31 @@ void ctrl_c_smash_handler(int sig_num){
 	cout << "smash: caught ctrl-C" << endl;
 	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
 }
-
-void ctrl_c_fg_handler(int sig_num){
+*/
+void ctrl_c_handler(int sig_num){
 	sigset_t mask_set;
 	sigset_t old_set;
 	sigfillset(&mask_set);
-	//sigprocmask(SIG_SETMASK, &mask_set, &old_set);
+	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
 	cout << "smash: caught ctrl-C" << endl;
-	kill(cur_pid, SIGKILL);
-	cout << "smash: process " << cur_pid << " was killed" <<endl;
-	//sigprocmask(SIG_SETMASK, &mask_set, &old_set);
+	if(cur_pid != smash_pid){
+		//int pid = getpid();
+		kill(cur_pid, SIGKILL);
+		cout << "smash: process " << cur_pid << " was killed" <<endl;
+	}
+	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
 
 }
 
-
-void ctrl_z_smash_handler(int sig_num){
+void ctrl_z_handler(int sig_num){
 	sigset_t mask_set;
 	sigset_t old_set;
 	sigfillset(&mask_set);
 	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
 	cout << "smash: caught ctrl-Z" << endl;
-	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
-}
-
-void ctrl_z_fg_handler(int sig_num){
-	sigset_t mask_set;
-	sigset_t old_set;
-	sigfillset(&mask_set);
-	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
-	cout << "smash: caught ctrl-Z" << endl;
-	kill(cur_pid, SIGSTOP);
-	cout << "smash: process " << cur_pid << " was stopped" <<endl;
+	if(cur_pid != smash_pid){
+		kill(cur_pid, SIGSTOP);
+		cout << "smash: process " << cur_pid << " was stopped" <<endl;
+	}
 	sigprocmask(SIG_SETMASK, &mask_set, &old_set);
 }
